@@ -1,40 +1,40 @@
 {
-  "name": "gulp-awslambda-3-status",
-  "version": "1.1.1",
-  "description": "A wrapper for checking the status of an AWS Lambda function",
-  "main": "index.js",
-  "scripts": {
-    "test": "node test/test.js",
-    "format": "eslint index.js",
-    "format-fix": "eslint --fix index.js"
-  },
-  "homepage": "https://github.com/netbymatt/gulp-awslambda-3-status",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/netbymatt/gulp-awslambda-3-status"
-  },
-  "bugs": {
-    "url": "https://github.com/netbymatt/gulp-awslambda-3-status/issues"
-  },
-  "keywords": [
-    "gulpplugin",
-    "aws",
-    "lambda",
-    "status"
-  ],
-  "engines": {
-    "node": ">=12.0.0"
-  },
-  "author": "Matt Walsh",
-  "license": "MIT",
-  "dependencies": {
-    "@aws-sdk/client-lambda": "^3.42.0",
-    "fancy-log": "^2.0.0"
-  },
-  "devDependencies": {
-    "eslint": "^8.3.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.3",
-    "gulp": "^4.0.2"
-  }
+	"name": "gulp-awslambda-3-status",
+	"version": "2.0.0",
+	"description": "A wrapper for checking the status of an AWS Lambda function",
+	"main": "index.mjs",
+	"scripts": {
+		"test": "gulp checkStatus",
+		"format": "eslint index.mjs",
+		"format-fix": "eslint --fix index.mjs"
+	},
+	"homepage": "https://github.com/netbymatt/gulp-awslambda-3-status",
+	"repository": {
+		"type": "git",
+		"url": "git://github.com/netbymatt/gulp-awslambda-3-status"
+	},
+	"bugs": {
+		"url": "https://github.com/netbymatt/gulp-awslambda-3-status/issues"
+	},
+	"keywords": [
+		"gulpplugin",
+		"aws",
+		"lambda",
+		"status"
+	],
+	"engines": {
+		"node": ">=12.0.0"
+	},
+	"author": "Matt Walsh",
+	"license": "MIT",
+	"dependencies": {
+		"@aws-sdk/client-lambda": "^3.588.0",
+		"fancy-log": "^2.0.0"
+	},
+	"devDependencies": {
+		"eslint": "^8.57.0",
+		"eslint-config-airbnb-base": "^15.0.0",
+		"eslint-plugin-import": "^2.29.1",
+		"gulp": "^5.0.0"
+	}
 }
